app/requirements.txt
Django==1.10.5
kafka-python==1.3.2
psycopg2==2.6.2
requests==2.13.0
